Bumped tensorflow==1.14.0 tensorflow-probability==0.7.0
PiperOrigin-RevId: 262154439

+++ configure.sh
@@ -21,15 +21,43 @@ function write_action_env_to_bazelrc() {
   write_to_bazelrc "build --action_env $1=\"$2\""
 }
 
-rm .bazelrc
-if python -c "import tensorflow" &> /dev/null; then
-    echo 'using installed tensorflow'
+# Remove .bazelrc if it already exist
+[ -e .bazelrc ] && rm .bazelrc
+
+# Check if it's installed
+if [[ $(pip show tensorflow) == *tensorflow* ]] || [[ $(pip show tf-nightly) == *tf-nightly* ]] ; then
+  echo 'Using installed tensorflow'
 else
-    pip install tensorflow
+  # Uninstall GPU version if it is installed.
+  if [[ $(pip show tensorflow-gpu) == *tensorflow-gpu* ]]; then
+    echo 'Already have gpu version of tensorflow installed. Uninstalling......\n'
+    pip uninstall tensorflow-gpu
+  elif [[ $(pip show tf-nightly-gpu) == *tf-nightly-gpu* ]]; then
+    echo 'Already have gpu version of tensorflow installed. Uninstalling......\n'
+    pip uninstall tf-nightly-gpu
+  fi
+  # Install CPU version
+  echo 'Installing tensorflow......\n'
+  pip install tensorflow
 fi
 
 TF_CFLAGS=( $(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_compile_flags()))') )
-TF_LFLAGS=( $(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_link_flags()))') )
+TF_LFLAGS="$(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_link_flags()))')"
+
+write_to_bazelrc "build --spawn_strategy=standalone"
+write_to_bazelrc "build --strategy=Genrule=standalone"
+write_to_bazelrc "build -c opt"
+
 
 write_action_env_to_bazelrc "TF_HEADER_DIR" ${TF_CFLAGS:2}
-write_action_env_to_bazelrc "TF_SHARED_LIBRARY_DIR" ${TF_LFLAGS:2}
+SHARED_LIBRARY_DIR=${TF_LFLAGS:2}
+SHARED_LIBRARY_NAME=$(echo $TF_LFLAGS | rev | cut -d":" -f1 | rev)
+if ! [[ $TF_LFLAGS =~ .*:.* ]]; then
+  if [[ "$(uname)" == "Darwin" ]]; then
+    SHARED_LIBRARY_NAME="libtensorflow_framework.dylib"
+  else
+    SHARED_LIBRARY_NAME="libtensorflow_framework.so"
+  fi
+fi
+write_action_env_to_bazelrc "TF_SHARED_LIBRARY_DIR" ${SHARED_LIBRARY_DIR}
+write_action_env_to_bazelrc "TF_SHARED_LIBRARY_NAME" ${SHARED_LIBRARY_NAME}
